Move Notifications package Turn on commented out "popup" for no devices..
